bugfixpy/tests/utils/test_validate.py
from unittest import TestCase

from ...utils.validate import has_valid_credentials


class TestValidate(TestCase):
    def test_is_valid_ticket_number(self) -> None:
        self.assertEqual()

    def test_is_valid_chlrq(self) -> None:
        self.assertEqual()

    def test_is_valid_chlc(self) -> None:
        self.assertEqual()

    def test_is_valid_challenge_id(self) -> None:
        self.assertEqual()

    def test_is_valid_fix_message(self) -> None:
        self.assertEqual()

    def test_is_valid_repository_name(self) -> None:
        self.assertEqual()

    def test_has_valid_credentials(self) -> None:
        self.assertTrue(has_valid_credentials())
